refactor : ChessGame 위치 변경
