+++ src/test/java/chess/domain/board/CatchPiecesTest.java
@@ -1,8 +1,8 @@
 package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
+import chess.domain.game.CatchPieces;
 import chess.domain.piece.Blank;
 import chess.domain.piece.King;
 import chess.domain.piece.Pawn;
