+++ src/test/java/chess/domain/game/ScoreTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.board.Board;
 import chess.domain.board.BoardFactory;
-import chess.domain.board.CatchPieces;
 import chess.domain.board.Position;
 import chess.domain.piece.Team;
 import org.junit.jupiter.api.BeforeEach;
